Move process logging to separate plugin
