+++ src/main.rs
@@ -1,10 +1,12 @@
 use std::time::Duration;
 
 use bevy::{app::ScheduleRunnerPlugin, prelude::*};
-use bevy_local_commands::{BevyLocalCommandsPlugin, ProcessError, ProcessOutput};
+use bevy_local_commands::BevyLocalCommandsPlugin;
 use engine::{EnginePlugin, StartEngine};
+use process_log::ProcessLogPlugin;
 
 mod engine;
+mod process_log;
 
 #[derive(Debug, Default, Component)]
 struct Game;
@@ -20,10 +22,10 @@ fn main() {
                 1.0 / 30.0,
             ))),
             BevyLocalCommandsPlugin,
+            ProcessLogPlugin,
             EnginePlugin,
         ))
         .add_systems(Startup, start_stockfish)
-        .add_systems(Update, (log_output, log_errors))
         .run();
 }
 
@@ -33,16 +35,4 @@ fn start_stockfish(mut commands: Commands, mut start_engine_event: EventWriter<S
         game_id,
         path: "stockfish".to_string(),
     });
-}
-
-fn log_output(mut output_event: EventReader<ProcessOutput>) {
-    for output in output_event.read() {
-        println!("{}", output.output.join("\n"));
-    }
-}
-
-fn log_errors(mut error_event: EventReader<ProcessError>) {
-    for error in error_event.read() {
-        eprintln!("{error:?}");
-    }
 }
